requirements.txt
tensorflow==2.13.0
keras==2.13.0
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
pillow==10.0.0
numpy==1.24.0
matplotlib==3.7.0
